requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
django-filter~=23.2
pip~=22.3.1
setuptools~=65.5.0